Fix weird slope
